requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-api-key==3.0.0
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
openapi==1.1.0
packaging==23.2
psycopg2-binary==2.9.9
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.1
rpds-py==0.13.2
sqlparse==0.4.4
typing-extensions==4.8.0
uritemplate==4.1.1
